refactor: move `decode_wal/2` back to `Message` module.
