+++ lib/postgrex_wal/message.ex
@@ -13,6 +13,7 @@ defmodule PostgrexWal.Message do
   """
 
   require Logger
+  alias PostgrexWal.StreamBoundaryError
 
   @modules %{
     Begin: ?B,
@@ -33,6 +34,9 @@ defmodule PostgrexWal.Message do
 
   @module_prefix PostgrexWal.Messages
   @streamable_modules [:Delete, :Insert, :Message, :Relation, :Truncate, :Type, :Update]
+  @stream_start_key @modules[:StreamStart]
+  @stream_stop_key @modules[:StreamStop]
+  @streamable_keys @modules |> Map.take(@streamable_modules) |> Map.values()
 
   expr =
     @modules
@@ -58,18 +62,38 @@ defmodule PostgrexWal.Message do
     end
   end
 
-  @spec __stream_start_key__() :: byte()
-  def __stream_start_key__, do: @modules[:StreamStart]
-
-  @spec __stream_stop_key__() :: byte()
-  def __stream_stop_key__, do: @modules[:StreamStop]
-
-  @spec __streamable_keys__() :: list(byte())
-  def __streamable_keys__, do: @modules |> Map.take(@streamable_modules) |> Map.values()
-
   @spec decode(event()) :: t()
   for {module, key} <- @modules do
     m = Module.concat(@module_prefix, module)
     def decode(<<unquote(key), payload::binary>>), do: unquote(m).decode(payload)
   end
+
+  @doc """
+  The logical replication protocol sends individual transactions one by one.
+  This means that all messages between a pair of Begin and Commit messages belong to the same transaction.
+  It also sends changes of large in-progress transactions between a pair of Stream Start and Stream Stop messages.
+  The last stream of such a transaction contains Stream Commit or Stream Abort message.
+  """
+
+  @type state() :: PostgrexWal.PgSource.t()
+  @spec decode_wal(event(), state()) :: {t(), state()}
+  def decode_wal(<<@stream_start_key, _rest::binary>> = event, state) do
+    if state.in_stream?, do: raise(StreamBoundaryError, "adjacent true")
+    {decode(event), %{state | in_stream?: true}}
+  end
+
+  def decode_wal(<<@stream_stop_key, _rest::binary>> = event, state) do
+    unless state.in_stream?, do: raise(StreamBoundaryError, "adjacent false")
+    {decode(event), %{state | in_stream?: false}}
+  end
+
+  def decode_wal(<<key, transaction_id::32, rest::binary>>, %{in_stream?: true} = state)
+      when key in @streamable_keys do
+    {
+      decode(<<key>> <> rest) |> struct!(transaction_id: transaction_id),
+      state
+    }
+  end
+
+  def decode_wal(event, state), do: {decode(event), state}
 end
